modify confirm_account_deletion
account cannot be deleted if user is owner of an organization.

I don't really like this here, as it opens up a can of worms...what
about the next dependent table that also wants to prevent users being
deleted.  What if I roll out just that, but not organizations?

etc etc

very chicken and egg